version 1: 重新设计页面
