+++ packages/blog/package.json
@@ -21,10 +21,10 @@
     "gitalk": "1.4.0",
     "highlight.js": "^9.12.0",
     "history": "^4.7.2",
-    "hoist-non-react-statics": "latest",
+    "hoist-non-react-statics": "3.1.0",
     "html-webpack-plugin": "2.29.0",
-    "immutable": "latest",
-    "invariant": "latest",
+    "immutable": "4.0.0-rc.12",
+    "invariant": "2.2.4",
     "lodash": "^4.17.10",
     "marked": "latest",
     "node-sass": "^4.7.2",
