+++ packages/blog/src/components/Card/card.scss
@@ -62,6 +62,7 @@
 }
 
 .card-title {
+
   span {
     font-size: 1.4rem;
     color: #555;
@@ -82,6 +83,10 @@
     }
   }
 
+  .title {
+    text-align: center;
+  }
+
   &:hover {
     cursor: pointer;
     span {
@@ -99,10 +104,10 @@
 }
 
 .card-content {
-  font-size: 14px;
+  font-size: 15px;
   padding: 1rem 0;
   span {
-    color: #a8a8a8;
+    color: $baseFontColor;
   }
 }
