+++ packages/blog/src/components/Header/assests/stylesheets/header.scss
@@ -10,7 +10,7 @@
   width: 100%;
   height: 52px;
   line-height: 52px;
-  background-color: #ffffff;
+  background-color: hsla(0,1%,100%,.7);
   border-bottom: 1px solid rgba(30, 35, 42, .06);
   -webkit-box-shadow: 0 1px 3px 0 rgba(0, 34, 77, .05);
   box-shadow: 0 1px 3px 0 rgba(0, 34, 77, .05);
