+++ packages/blog/src/components/Header/header.js
@@ -8,7 +8,7 @@ import './assests/stylesheets/header.scss';
 const menu = [
     {
         name: '文章',
-        path: '/articles',
+        path: '/',
     },
     {
         name: '归档',
@@ -85,7 +85,7 @@ class Header extends React.Component{
 
                     <div className='header-menu'>
                         {
-                            menu.map(v =>  <Link className={classNames('header-path-link', {[`active`]: pathname.includes(`${v.path}`)})}
+                            menu.map(v =>  <Link className={classNames('header-path-link', {[`active`]: pathname === `${v.path}`})}
                                                  to={{ pathname: `${v.path}`}} key={v.path}>{v.name}</Link>)
                         }
                     </div>
@@ -94,7 +94,14 @@ class Header extends React.Component{
                         <div className='header-nav-main'>
                             <ul>
                                 {
-                                    menu.map(v => <li className='nav-item' key={v.path}><Link to={{ pathname: `${v.path}`}}>{v.name}</Link></li>)
+                                    menu.map(v => <li className='nav-item' key={v.path}>
+                                        <Link
+                                            className={classNames('header-path-link', {[`active`]: pathname === `${v.path}`})}
+                                            to={{ pathname: `${v.path}`}}
+                                        >
+                                            {v.name}
+                                        </Link>
+                                    </li>)
                                 }
                             </ul>
                         </div>
