--- packages/blog/src/routes/App/constants.js
@@ -1,3 +1,0 @@
-export const APP = 'app';
-
-export const APP_PAGE = 'APP_PAGE';
