--- packages/blog/src/routes/App/containers.js
@@ -1,17 +1,0 @@
-import {connect} from 'react-redux';
-
-import * as App from './constants';
-import {changeAppPage} from "./modules";
-
-const mapDispatchToProps = {
-    changeAppPage: (appPage) => changeAppPage(appPage),
-
-};
-
-const mapStateToProps = (state) => {
-    return {
-        appPage: state.get(App.APP).get(App.APP_PAGE),
-    };
-};
-
-export const withConnect = connect(mapStateToProps, mapDispatchToProps);
