--- packages/blog/src/routes/App/index.js
@@ -1,11 +1,0 @@
-import { compose } from 'redux';
-
-import injectReducer from '../../store/reducers';
-import App from './app';
-import { withConnect } from './containers';
-import reducer from './modules';
-
-const withReducer = injectReducer({key: 'app', reducer});
-
-export default compose(withReducer, withConnect)(App);
-
