--- packages/blog/src/routes/Articles/constants.js
@@ -1,4 +1,0 @@
-export const ARTICLES = 'articles';
-
-export const ARTICLES_DATA = 'ARTICLES_DATA';
-export const CLEAR_RELATIVE = 'CLEAR_RELATIVE';
