--- packages/blog/src/routes/Articles/stylesheets/articles.scss
@@ -1,13 +1,0 @@
-.articles {
-  width: 80%;
-
-  h1.tag-name {
-    color: $baseFontColor;
-    margin-bottom: 1rem;
-    font-size: 1.6rem;
-  }
-
-  @media only screen and (max-width: 768px) {
-    width: 100%;
-  }
-}
