--- packages/blog/src/routes/Home/stylesheets/home.scss
@@ -1,95 +1,0 @@
-.home {
-  height: 100%;
-  width: 100%;
-  //background-image: linear-gradient(to top, #e6e9f0 0%, #eef1f5 100%);
-  background-image: linear-gradient(120deg, #fdfbfb 0%, #ebedee 100%);
-}
-
-.home {
-  display: flex;
-  display: -webkit-flex;
-  display: -moz-flex;
-  flex-flow: column;
-  justify-content: center;
-  text-align: center;
-
-  .home-text {
-    padding: 10px;
-    font-size: 1.4rem;
-    font-weight: 500;
-  }
-
-  .home-text .type {
-    display: inline;
-  }
-
-  .home-text .chapter {
-    font-size: .8rem;
-    font-weight: 200;
-    text-align: right;
-    padding: 1rem 1rem 0 0;
-  }
-  .typed-cursor {
-    font-weight: 500;
-  }
-
-  .blog-button {
-
-    button {
-      width: 100px;
-      max-width: 100px;
-      background: none;
-      padding: 8px;
-      border-radius: 2px;
-      font-size: 1rem;
-      font-weight: 300;
-      border: 1px solid $baseFontColor;
-      margin: 1rem 2rem;
-      outline: none;
-
-      &:hover {
-        cursor: pointer;
-      }
-    }
-  }
-}
-
-//.home {
-//  -webkit-animation: animatename 1s;
-//  -moz-animation: animatename 1s;
-//  animation: animatename 1s;
-//}
-//@keyframes animatename {
-//  0% {
-//    opacity: 0;
-//    -webkit-transform: translateY(-100%);
-//    -ms-transform: translateY(-100%);
-//    transform: translateY(-100%);
-//  }
-//
-//  100% {
-//    opacity: 1;
-//    -webkit-transform: translateY(0%);
-//    -ms-transform: translateY(0%);
-//    transform: translateY(0%);
-//  }
-//}
-//
-//@keyframes animatename-reverse {
-//
-//  100% {
-//    opacity: 1;
-//    -webkit-transform: translateY(0%);
-//    -ms-transform: translateY(0%);
-//    transform: translateY(0%);
-//  }
-//
-//  0% {
-//    opacity: 0;
-//    -webkit-transform: translateY(-100%);
-//    -ms-transform: translateY(-100%);
-//    transform: translateY(-100%);
-//  }
-//}
-
-
