+++ packages/blog/src/stylesheets/index.scss
@@ -18,8 +18,6 @@ body {
 .container {
   flex: 1 0 auto;
   margin: 4.5rem auto;
-  width: 90%;
-  max-width: 1024px;
   display: flex;
   display: -webkit-flex;
   display: -moz-flex;
